完善设备 酒 提成 角色 用户 区域相关代码

+++ src/main/java/com/hzxy/modules/sellwine/VO/WinePercentageVO.java
@@ -14,6 +14,6 @@ public class WinePercentageVO {
 
     private String roleName;
 
-    private Integer percentage;
+    private Double percentage;
 
 }
